app: add handler to get the cost of an opal train trip
You will need to refresh your dependencies (pip3 install -r
requirements.txt), because of the added DistanceMatrix dep.

Currently this is a proof of concept, I plan to add support for
different ticket types soon, probably will use a dictionary? At the moment everything is calculated for an adult at peek hour.

The main drawback for using this approach is the request limit, which is
a meagre 2500 a day... I suppose we could hard cache everything, because
realistically train stations are not going to change very often at all.