js: use incremental fares
a->b->c on one mode will be treated as a single trip from a->c.
Note that this doesn't account for going backwards. AFAIK what actually
happens is that if you go from a->b->c, where b is the furthest from a
you go, the resultant fare will just be from a->b.